Preparing EventSummary (used in API) in EventService
