+++ src/main/java/pl/net/rogala/eventy/api/ApiEventController.java
@@ -6,26 +6,27 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import pl.net.rogala.eventy.service.EventService;
 
 import java.util.List;
 
 @RestController
 @RequestMapping("/api")
 public class ApiEventController {
-    private ApiEventService apiEventService;
+    private EventService eventService;
 
     @Autowired
-    public ApiEventController(ApiEventService apiEventService) {
-        this.apiEventService = apiEventService;
+    public ApiEventController(EventService eventService) {
+        this.eventService = eventService;
     }
 
     @GetMapping("/events")
-    public ResponseEntity<List<EventSummary>> showAllEvents(){
-        return ResponseEntity.ok(apiEventService.collectAllEvents());
+    public ResponseEntity<List<EventSummary>> showAllEvents() {
+        return ResponseEntity.ok(eventService.collectAllEvents());
     }
 
     @GetMapping("/events/{startDate},{stopDate}")
-    public ResponseEntity<List<EventSummary>> showEventsByDateRange(@PathVariable String startDate, @PathVariable String stopDate){
-        return ResponseEntity.ok(apiEventService.collectEventsByDateRange(startDate, stopDate));
+    public ResponseEntity<List<EventSummary>> showEventsByDateRange(@PathVariable String startDate, @PathVariable String stopDate) {
+        return ResponseEntity.ok(eventService.collectEventsByDateRange(startDate, stopDate));
     }
 }
