+++ src/main/java/pl/net/rogala/eventy/service/EventService.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
+import pl.net.rogala.eventy.api.EventSummary;
 import pl.net.rogala.eventy.converter.EventConverter;
 import pl.net.rogala.eventy.entity.Comment;
 import pl.net.rogala.eventy.entity.Event;
@@ -12,8 +13,10 @@ import pl.net.rogala.eventy.repository.EventRepository;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class EventService {
@@ -69,5 +72,22 @@ public class EventService {
 
     public List<Event> findEventsByDateRange(LocalDate startDate, LocalDate stopDate) {
         return eventRepository.findAllByStartDateAfterAndStopDateBefore(startDate, stopDate);
+    }
+
+    public List<EventSummary> collectAllEvents() {
+        List<EventSummary> eventSummaryList = showEventList().stream()
+                .map(event -> new EventSummary(event.getId(), event.getName(), event.getStartDate(), event.getStopDate()))
+                .collect(Collectors.toList());
+        return eventSummaryList;
+    }
+
+    public List<EventSummary> collectEventsByDateRange(String startDate, String stopDate) {
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+        LocalDate start = LocalDate.parse(startDate, formatter);
+        LocalDate stop = LocalDate.parse(stopDate, formatter);
+        List<EventSummary> eventSummaryListWithDateRange = findEventsByDateRange(start, stop).stream()
+                .map(event -> new EventSummary(event.getId(), event.getName(), event.getStartDate(), event.getStopDate()))
+                .collect(Collectors.toList());
+        return eventSummaryListWithDateRange;
     }
 }
